Move invoice code to InvoiceController
